apps/pomodoro_timer/pyproject.toml
[project]
name = "pomodoro-timer"
version = "0.1.0"
description = "A minimal Python CLI tool for Pomodoro technique focus sessions with beautiful terminal output"
requires-python = ">=3.12"
dependencies = [
    "plyer>=2.1.0",  # Used for Windows notifications only; macOS/Linux use native OS commands
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.20.1",
]
